+++ src/CleanRegex/Match/Stream.php
@@ -73,9 +73,9 @@ class Stream implements \Countable, \IteratorAggregate
     public function first(callable $consumer = null)
     {
         if ($consumer === null) {
-            return $this->firstOptional()->orThrow();
+            return $this->firstOptional()->get();
         }
-        return $this->firstOptional()->map($consumer)->orThrow();
+        return $this->firstOptional()->map($consumer)->get();
     }
 
     public function findFirst(callable $consumer): Optional
@@ -100,7 +100,7 @@ class Stream implements \Countable, \IteratorAggregate
 
     public function nth(int $index)
     {
-        return $this->findNth($index)->orThrow();
+        return $this->findNth($index)->get();
     }
 
     public function findNth(int $index): Optional